Change runners API (#38)
* Remove runEvolution and getGenerationsIterator

* Rename Genemo.runEvolutionAsync to Genemo.run

* Remove synchronous reproduce function
