+++ src/index.js
@@ -1,11 +1,11 @@
-const { runEvolution, runEvolutionAsync, getGenerationsIterator } = require('./runners');
+const { run } = require('./runners');
 
 const selection = require('./selection');
 const mutation = require('./mutation');
 const crossover = require('./crossover');
 
 const { generateInitialPopulation } = require('./generateInitialPopulation');
-const { reproduce, reproduceAsync } = require('./reproduce');
+const { reproduce } = require('./reproduce');
 const { stopCondition } = require('./stopConditions');
 const randomSequenceOf = require('./randomSequenceOf');
 const randomPermutationOf = require('./randomPermutationOf');
@@ -13,9 +13,7 @@ const elitism = require('./elitism');
 const logIterationData = require('./logIterationData');
 
 module.exports = {
-  runEvolution,
-  runEvolutionAsync,
-  getGenerationsIterator,
+  run,
 
   selection,
   mutation,
@@ -23,7 +21,6 @@ module.exports = {
 
   generateInitialPopulation,
   reproduce,
-  reproduceAsync,
   stopCondition,
   logIterationData,
   randomSequenceOf,
