+++ test/reproduce.test.js
@@ -1,8 +1,9 @@
 const Genemo = require('../src');
+const asyncify = require('../src/utils/asyncify');
 const cyclicProvider = require('./test-utils/cyclicProvider');
 
 describe('reproduce', () => {
-  test('Returns correct population with odd number of individuals', () => {
+  test('Returns correct population with odd number of individuals', async (done) => {
     const random = cyclicProvider([
       0.41, 0.21,
       0.0, 0.99,
@@ -22,11 +23,12 @@ describe('reproduce', () => {
       mutate: individual => individual,
       crossover: (([a, b]) => [a, b]),
     });
-    const result = reproduce(evaluatedPopulation, random);
+    const result = await reproduce(evaluatedPopulation, random);
     expect(result).toStrictEqual(expectedResult);
+    done();
   });
 
-  test('Calls mutation for the right individuals', () => {
+  test('Calls mutation for the right individuals', async (done) => {
     const random = cyclicProvider([
       0.41, 0.21,
       0.0, 0.99,
@@ -44,11 +46,12 @@ describe('reproduce', () => {
     const expectedResult = [2, 101, 0, 104, 1];
 
     const reproduce = Genemo.reproduce({
-      mutate: individual => individual + 100,
-      crossover: (([a, b]) => [a, b]),
+      mutate: asyncify(individual => individual + 100),
+      crossover: asyncify(([a, b]) => [a, b]),
       mutationProbability: 0.1,
     });
-    const result = reproduce(evaluatedPopulation, random);
+    const result = await reproduce(evaluatedPopulation, random);
     expect(result).toStrictEqual(expectedResult);
+    done();
   });
 });
